[BUGFIX] Check if file exists before deletion
File deletion in "processed" folder no longer throws an exception when
trying to delete a file that does not exist.

Resolves: #67432
Releases: master, 6.2
Reviewed-on: http://review.typo3.org/40238
Tested-by: Michael Oehlhof <typo3@oehlhof.de>
Tested-by: Alexander Opitz <opitz.alexander@googlemail.com>

+++ typo3/sysext/core/Classes/Resource/ResourceStorage.php
@@ -1560,9 +1560,11 @@ class ResourceStorage implements ResourceStorageInterface {
 
 		$this->emitPreFileDeleteSignal($fileObject);
 
-		$result = $this->driver->deleteFile($fileObject->getIdentifier());
-		if ($result === FALSE) {
-			throw new Exception\FileOperationErrorException('Deleting the file "' . $fileObject->getIdentifier() . '\' failed.', 1329831691);
+		if ($this->driver->fileExists($fileObject->getIdentifier())) {
+			$result = $this->driver->deleteFile($fileObject->getIdentifier());
+			if (!$result) {
+				throw new Exception\FileOperationErrorException('Deleting the file "' . $fileObject->getIdentifier() . '\' failed.', 1329831691);
+			}
 		}
 		// Mark the file object as deleted
 		if ($fileObject instanceof File) {
